requirements.txt
numpy>=1.24.2
pandas>=1.5.3
yfinance>=0.2.18
python-binance>=1.0.17
scikit-learn>=1.2.2
torch>=2.0.0
ta>=0.10.2
python-telegram-bot>=20.0
matplotlib>=3.7.1
seaborn>=0.12.2
requests>=2.28.2
python-dotenv>=1.0.0
SQLAlchemy>=1.4.23
Flask>=2.0.1
dash>=2.0.0
plotly>=5.13.1
protobuf>=3.20.0 